feat: refactor TicketService package structure and update imports
